Moved sqrtm_psd → utils.py
